backend/requirements.txt
fastapi
uvicorn
pdfminer.six
pdfkit
python-multipart
fastapi-mail
jose
passlib[bcrypt]
sqlmodel